WIP: lotsa stuff - bye yarn - moves code out of giant index.js into smaller files - adds tests/bir-vue (Vue CLI vanilla) for local testing - updates CONTRIBUTING - adds LICENSE
