+++ tests/bir-vue/vue.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  chainWebpack: (config) => {
+    config.resolve.symlinks(false)
+  }
+}
